+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 Changelog
 =========
 
+20.06
+-----
+
+* Added the DND and forwards status synchronization for Yealink phones
+* Added the following endpoints:
+
+  * GET `/0.1/yealink/users/<uuid>/services/dnd/enable`
+  * GET `/0.1/yealink/users/<uuid>/services/dnd/disable`
+  * GET `/0.1/yealink/directories/lookup/<profile>`
+
 20.03
 -----
 
 * Added a plugin to handle BLF events.
-
 
 19.17
 -----